+++ app/controllers/admin/cycles/plugin_relations/petitions_controller.rb
@@ -35,6 +35,7 @@ class Admin::Cycles::PluginRelations::PetitionsController < Admin::ApplicationCo
 
     response = detail_updater.perform @petition, petition_params, petition_body
     if response.success
+      shared_link_generation response
       enqueue_pdf_generation response
       flash[:success] = "Projeto de Lei salvo com sucesso."
       redirect_to [:admin, @cycle, @plugin_relation, :petitions]
@@ -75,6 +76,10 @@ class Admin::Cycles::PluginRelations::PetitionsController < Admin::ApplicationCo
 
   def enqueue_pdf_generation(use_case_response)
     PetitionPdfGenerationWorker.perform_async id: use_case_response.version.id if use_case_response.version
+  end
+
+  def shared_link_generation(response)
+    PetitionShareLinkGenerationWorker.perform_async id: response.detail.id
   end
 
   def detail_updater