+++ db/migrate/20180502131815_add_share_link_to_petition_plugin_details.rb
@@ -1,0 +1,5 @@
+class AddShareLinkToPetitionPluginDetails < ActiveRecord::Migration
+  def change
+    add_column :petition_plugin_details, :share_link, :string
+  end
+end